Deepvue/GithubApiChart/requirements.txt
django
requests
django-hosts
djangorestframework
django-cors-headers
